Multi-arch each of the builds
Had to drop the yarn testing in the app build because arm wouldn't
install the fsevents package without a LOT of extra deps (python,
make, gcc, etc.). So, skipping that for now. We can do the test
as part of the verify stage and do it only in a single arch, as
that code should run the same everywhere.